api/requirements.txt
fastapi
uvicorn
supabase
requests
sentence-transformers
torch --index-url https://download.pytorch.org/whl/cpu
python-dotenv
pydantic